Several code enhancements Make graph-of-reachable-objects always calculate shortest paths to every object from the root set of objects.
Add view* write-drawing-file* write-dot-file* functions that are
identical to the ones with names not ending in *, except they take a
collection of objects from which to generate a graph, not only one.

Add helper functions to cljol.graph namespace to make it easy to
filter the nodes by shortest path distance, or to remove all leaf
nodes.