Use JOL's calculated node distance instead of Ubergraph shortest path
Before making this change, I checked that a graph of about 150,000
nodes and 300,000+ edges had the same distances calculated via both
methods.

This will not speed things up much, but it does a little bit.  The
shortest path function is still there in case someone wants to
calculate shortest paths starting from a set of nodes that was not the
original root nodes used to create the graph.