Add :bounded2 version of calculating reachable # nodes & total size
It is currently even slower than :bounded, and I am not sure why yet
it would be, but hopefully can be optimized to be faster than :bounded
without a lot of work.

Added new namespace containing my-time macro and print-perf-stats, so
it is easier to use from multiple other namespaces.