Added :bounded3 option for calculating total node sizes
For some reason I do not understand yet, on a single Var in user
namespace, :bounded takes about 20 sec, but :bounded3 takes about 300
sec.  One significant difference between these two is that :bounded3
runs on scc-graph, which has sets of integers for nodes, not integers
as :bounded has.

After committing this code, I may try an experiment in which :bounded3
uses an scc-graph with integers for nodes, with a separate map from
those integers to the original sets of integers that they are now.