Update :bounded3 code so it ran correctly again.
Tested that it is pretty much identical in speed to latest :bounded4,
so both are pleasingly fast compared to what they were before I wrote
a faster pre-traverse function than Loom v1.0.1's pre-traverse.

The only difference between them is :bounded4 creates an scc-graph
with integer values for nodes, whereas :bounded3 is the older one that
uses sets of integers for nodes.  Apparently the fact that some of
these sets can be large is not a performance problem, as I originally
guessed/hoped they would not be.  Thus I will likely delete the
:bounded4 code as an unnecessary experiment soon.