Experiment with a cljol.object-walk Clojure object walker
instead of GraphLayout/parseInstace.  Initial results show that the
object walking part takes about 2x more time than
GraphLayout/parseInstance.

Made reachable-objmaps use mapv, so it is eager, and thus using
my-time to measure its time counts all of the time to create the
objmaps.  It was not including that time before due to lazy map.